fork check

+++ app/src/main/java/com/eyetracker/mobile/ui/camera/CameraScreen.java
@@ -7,7 +7,7 @@ import com.eyetracker.mobile.model.Frame;
  */
 public interface CameraScreen {
 
-    void showProcessedImage(byte[] image);
+    void showProcessedImage(int[] image);
 
     void discardResults();
